Move remarshaling to happen only during comparison, instead of manifest generation
